refactor(cache): 스키마 정의 파일 분리 및 재사용
